+++ app/src/main/java/com/example/cyclosens/classes/Activity.java
@@ -11,8 +11,8 @@ public class Activity  implements Serializable {
     private long duration;
     private int bpmAv;
     private int strengthAv;
-    private Float speedAv;
-    private Float distance;
+    private int speedAv;
+    private int distance;
     private ArrayList<Position> positionList;
 
     public String getNameActivity() { return nameActivity; }
@@ -21,8 +21,8 @@ public class Activity  implements Serializable {
     public String getKey() { return key; }
     public int getBpmAv() { return bpmAv; }
     public int getStrengthAv() { return strengthAv; }
-    public Float getSpeedAv() { return speedAv; }
-    public Float getDistance() { return distance; }
+    public int getSpeedAv() { return speedAv; }
+    public int getDistance() { return distance; }
     public ArrayList<Position> getPositionList() { return positionList; }
 
     public void setNameActivity(String nameActivity) { this.nameActivity = nameActivity; }
@@ -31,11 +31,11 @@ public class Activity  implements Serializable {
     public void setKey(String key) { this.key = key; }
     public void setBpmAv(int bpmAv) { this.bpmAv = bpmAv; }
     public void setStrengthAv(int strengthAv) { this.strengthAv = strengthAv; }
-    public void setSpeedAv(Float speedAv) { this.speedAv = speedAv; }
-    public void setDistance(Float distance) { this.distance = distance; }
+    public void setSpeedAv(int speedAv) { this.speedAv = speedAv; }
+    public void setDistance(int distance) { this.distance = distance; }
     public void setPositionList(ArrayList<Position> positionList) { this.positionList = positionList; }
 
-    public Activity(String key, String nameActivity, String dateActivity, long duration, int bpmAv, int strengthAv, Float speedAv, Float distance, ArrayList<Position> positionList) {
+    public Activity(String key, String nameActivity, String dateActivity, long duration, int bpmAv, int strengthAv, int speedAv, int distance, ArrayList<Position> positionList) {
         setKey(key);
         setNameActivity(nameActivity);
         setDateActivity(dateActivity);
